# public/openapi.yaml
openapi: 3.0.3
info:
  title: OfferGuide API
  version: 2.0.0
  description: >
    OfferGuide (sanjeeda.io career services portal) backend contract.


    Sprint 8 completed this file. Sprints 1–7 shipped 34 paths but only
    Sprint 5's own 13 were ever documented here, which made Story 8.1.1
    ("verify every endpoint in Swagger") impossible to execute. Every route
    that exists in `src/app/api/offerguide/` is now declared, with its auth
    mode, status codes and response shape.


    ## Auth modes

    Four modes, named exactly as the Sprint 8 handoff §3.1 table names them:


    * **PUBLIC** — no cookie, no token. Reference data only.

    * **guestOrAuth** — a `guestToken` httpOnly cookie **or** a portal JWT.
      Behaviour is identical either way, scoped to whichever identity is
      present; a JWT wins when both are sent. Never returns one candidate's
      data to another.

    * **bearerAuth** — a portal JWT only. A guest cookie is rejected.

    * **adminAuth** — the interim `x-og-admin-token` gate. Rejects no
      credentials, a guest cookie, and an ordinary registered-user JWT alike.

    ### Swagger UI limitation (read before verifying)

    The **Authorize** button can only drive the JWT and admin-token cases. It
    cannot attach the `guestToken` httpOnly cookie. Every `guestOrAuth`
    operation must therefore be verified **twice**: once here with a JWT, and
    once outside Swagger (browser or `curl -b`) with only the guest cookie
...


...

...

...

...

...


...

...


...

...
servers:
...
security: []
paths:
...
  /api/offerguide/candidate-profile:
    get:
      tags: [Candidate Profile]
      summary: Get the caller's candidate profile
      description: guestOrAuth. Scoped to the caller's own identity.
      security: [{ BearerAuth: [] }, { GuestCookie: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CandidateProfile" }
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Candidate Profile]
      summary: Create the caller's candidate profile
      description: >
        guestOrAuth. With no identity present, mints a `guestToken` and sets
        it as an httpOnly cookie on the response — this is the operation that
        makes a guest a guest.
      security: [{ BearerAuth: [] }, { GuestCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CandidateProfile" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CandidateProfile" }
        "400":
          $ref: "#/components/responses/BadRequest"
    patch:
      tags: [Candidate Profile]
      summary: Update the caller's candidate profile
      description: >
        guestOrAuth. Partial merge — only the fields sent are changed. A
        return visit updates the existing profile rather than creating a
        second one.
      security: [{ BearerAuth: [] }, { GuestCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CandidateProfile" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CandidateProfile" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/offerguide/admin/config/geography/{countryCode}/cities:
    parameters: [{ $ref: "#/components/parameters/CountryCode" }]
    post:
      tags: [Admin Config]
      summary: Add a city to a country
      description: >
        adminAuth. Cities are embedded in their country document, so this is a
        sub-resource POST rather than a create against a collection of its
        own. `cityId` is unique within its country; a duplicate returns 409.
        Returns the whole updated country.
      security: [{ AdminToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GeographyCity" }
      responses:
        "201":
          description: Created — returns the updated country document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Geography" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: A city with that `cityId` already exists in this country
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        The portal's own JWT. On a `guestOrAuth` route it takes precedence
        over the guest cookie when both are present.
    GuestCookie:
      type: apiKey
      in: cookie
      name: guestToken
      description: >
        httpOnly cookie minted by `POST /candidate-profile` when no identity
        is present. **Swagger UI cannot attach this** — verify guest-mode
        behaviour outside Swagger with `curl -b` or a browser session.
    AdminToken:
      type: apiKey
      in: header
      name: x-og-admin-token
      description: >
        *** TEMPORARY, Sprint 9 placeholder for real RBAC *** — a single
        env-configured token (`OFFERGUIDE_ADMIN_TOKEN`), not a role column,
        not `usePermission()`. No `role` field is referenced anywhere in the
        gate's logic, because it does not exist yet. Fails closed: an unset
        env var means no request can pass. Every `/admin/config/*` route
        requires it, and it rejects an ordinary registered-user JWT — a
        registered user is not an admin.
  parameters:
...
    CountryCode:
      name: countryCode
      in: path
      required: true
      schema: { type: string }
      example: PK
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid credentials for this operation's auth mode
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: >
        Not found — also returned when the resource belongs to another
        candidate, so the contract never confirms that someone else's data
        exists.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }

    # ----------------------------------------------------------- candidate
    ConsentSettings:
      type: object
      description: >
        Candidate-level, stored on `candidate_profiles.consentSettings`.
        Every toggle defaults to **Off** — privacy-first. When
        `shareAnonymous` is false, every selection is treated as Off
        regardless of its stored value, and the sub-toggles are dimmed in the
        UI. Framing is anonymised throughout: no copy implies personal data is
        shared.
      properties:
        shareAnonymous:
          type: boolean
          default: false
          description: The master toggle, `consent_share_anonymous`.
        selections:
          type: object
          additionalProperties: { type: boolean }
          description: >
            Keyed by `toggleId`. Validated against `OgConsentToggles`, so an
            unknown or retired id is rejected with 400.
    CandidateProfile:
      type: object
      description: >
        SCR-001. 24 fields, of which only `careerStage` and
        `preferredWorkArrangement` are required. One profile per identity;
        a return visit updates it rather than creating a second.
      properties:
        id: { type: integer, readOnly: true }
        userInfoId:
          type: integer
          nullable: true
          description: Null while the candidate is a guest.
        guestToken:
          type: string
          nullable: true
          readOnly: true
        careerStage: { type: string }
        preferredWorkArrangement: { type: string }
        currentCountry: { type: string, nullable: true }
        currentCity: { type: string, nullable: true }
        consentSettings: { $ref: "#/components/schemas/ConsentSettings" }
      additionalProperties: true

    # --------------------------------------------------------------- session
...
    GeographyCity:
      type: object
      required: [cityId, name]
      properties:
        cityId: { type: string }
        name: { type: string }
        active: { type: boolean, default: true }
    Geography:
      type: object
      required: [countryCode, countryName]
      properties:
        countryCode: { type: string, example: PK }
        countryName: { type: string, example: Pakistan }
        active: { type: boolean, default: true }
        cities:
          type: array
          items: { $ref: "#/components/schemas/GeographyCity" }
...